+++ src/components/utils/connectToAdServer.js
@@ -3,6 +3,8 @@ import AdvertisingContext from '../../AdvertisingContext';
 
 export default Component => props => (
     <AdvertisingContext.Consumer>
-        {({ activate, active }) => <Component {...props} activate={activate} active={active} />}
+        {({ activate, active, shouldRefresh }) => (
+            <Component {...props} activate={activate} active={active} shouldRefresh={shouldRefresh} />
+        )}
     </AdvertisingContext.Consumer>
 );